+++ migrations/versions/665569c27179_make_password_hash_null.py
@@ -5,17 +5,15 @@ Revises: a7076e5f1905
 Create Date: 2026-03-24 09:13:18.927467
 
 """
-from typing import Sequence, Union
+from collections.abc import Sequence
 
-from alembic import op
-import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
 revision: str = '665569c27179'
-down_revision: Union[str, Sequence[str], None] = 'a7076e5f1905'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
+down_revision: str | Sequence[str] | None = 'a7076e5f1905'
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
 
 
 def upgrade() -> None: