+++ src/animus/database/sqlalchemy/models/intake/__init__.py
@@ -1,5 +1,11 @@
 from .analysis_model import AnalysisModel
 from .petition_model import PetitionModel
 from .petition_summary_model import PetitionSummaryModel
+from .precedent_model import PrecedentModel
 
-__all__ = ['AnalysisModel', 'PetitionModel', 'PetitionSummaryModel']
+__all__ = [
+    'AnalysisModel',
+    'PetitionModel',
+    'PetitionSummaryModel',
+    'PrecedentModel',
+]