+++ src/animus/database/sqlalchemy/repositories/intake/__init__.py
@@ -3,9 +3,11 @@ from .sqlalchemy_petition_summaries_repository import (
     SqlalchemyPetitionSummariesRepository,
 )
 from .sqlalchemy_petitions_repository import SqlalchemyPetitionsRepository
+from .sqlalchemy_precendents_repository import SqlalchemyPrecedentsRepository
 
 __all__ = [
     'SqlalchemyAnalisysesRepository',
     'SqlalchemyPetitionsRepository',
     'SqlalchemyPetitionSummariesRepository',
+    'SqlalchemyPrecedentsRepository',
 ]