+++ src/animus/pubsub/inngest/jobs/intake/__init__.py
@@ -1,0 +1,3 @@
+from .vectorize_precedents_job import VectorizePrecedentsJob
+
+__all__ = ['VectorizePrecedentsJob']