+++ src/animus/rest/pangea/services/mappers/__init__.py
@@ -1,0 +1,3 @@
+from .pangea_bnp_precedent_mapper import PangeaBnpPrecedentMapper
+
+__all__ = ['PangeaBnpPrecedentMapper']