+++ src/animus/rest/pangea/services/models/__init__.py
@@ -1,0 +1,3 @@
+from .pangea_bnp_precedent import PangeaBnpResponse
+
+__all__ = ['PangeaBnpResponse']